feat(visor): serve the standalone wasm-visor / testing harness from the visor process
`skywire cli hv serve` builds the standalone wasm-visor PWA (+ optional
/ctl/* testing harness) entirely from the running binary's embedded wasm
+ UI. This lifts that surface into the visor itself so it can be hosted by
the one visor process instead of a second command — set
`hypervisor.wasm_serve.addr` (e.g. ":8443") in the config and the visor
serves it alongside the HV UI; an operator's rebuild-restart loop then
re-embeds + re-serves the latest wasm every restart, with nothing extra to
launch.

- Extract the serve handler builder into pkg/visor.ServeWasm (+
  WasmServeConfig) with the helpers it needs (injectWasmBoot, the
  self-signed localhost cert, the wallet dmsg-fetch shim, the
  access-password gate). It's ctx-driven and self-contained — no visor RPC
  dependency — so it still serves standalone.
- Add hypervisor.wasm_serve config (addr/tls/harness/no_wallet/variant/
  password); the visor starts ServeWasm from it (off unless addr is set).
- Rewrite `hv serve` to call the shared ServeWasm (removes ~450 lines of
  duplicated handler/helper code that would otherwise drift).

Standalone `skywire cli hv serve --tls --harness --addr :8443` is
unchanged. Build + vet + visorconfig tests green.
